Separate LinkableElementProperties.
