+++ metricflow/model/semantics/linkable_spec_resolver.py
@@ -4,13 +4,13 @@ import logging
 import time
 from collections import defaultdict
 from dataclasses import dataclass
-from enum import Enum
 from typing import Tuple, Sequence, Dict, List, Optional, FrozenSet
 
 from metricflow.model.objects.data_source import DataSource
 from metricflow.model.objects.elements.dimension import DimensionType, Dimension
 from metricflow.model.objects.elements.identifier import IdentifierType
 from metricflow.model.objects.user_configured_model import UserConfiguredModel
+from metricflow.model.semantics.linkable_element_properties import LinkableElementProperties
 from metricflow.object_utils import pformat_big_objects, flatten_nested_sequence
 from metricflow.references import MeasureReference, MetricReference
 from metricflow.specs import (
@@ -24,41 +24,6 @@ from metricflow.specs import (
 from metricflow.time.time_granularity import TimeGranularity
 
 logger = logging.getLogger(__name__)
-
-
-class LinkableElementProperties(Enum):
-    """The properties associated with a valid linkable element.
-
-    Local means an element that is defined within the same data source as the measure. This definition is used
-    throughout the related classes.
-    """
-
-    # A local element as per above definition.
-    LOCAL = "local"
-    # A local dimension that is prefixed with a local primary identifier.
-    LOCAL_LINKED = "local_linked"
-    # An element that was joined to the measure data source by an identifier.
-    JOINED = "joined"
-    # An element that was joined to the measure data source by joining multiple data sources.
-    MULTI_HOP = "multi_hop"
-    # A time dimension that is a version of a time dimension in a data source, but at a different granularity.
-    DERIVED_TIME_GRANULARITY = "derived_time_granularity"
-    # Refers to an identifier, not a dimension.
-    IDENTIFIER = "identifier"
-    # After an intersection operation.
-    INTERSECTED = "intersected"
-
-    @staticmethod
-    def all_properties() -> FrozenSet[LinkableElementProperties]:  # noqa: D
-        return frozenset(
-            {
-                LinkableElementProperties.LOCAL,
-                LinkableElementProperties.LOCAL_LINKED,
-                LinkableElementProperties.JOINED,
-                LinkableElementProperties.MULTI_HOP,
-                LinkableElementProperties.DERIVED_TIME_GRANULARITY,
-            }
-        )
 
 
 @dataclass(frozen=True)
